requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pinecone-client[grpc]==2.2.1
openai==0.27.0
python-dotenv==0.19.0